accuracy rating generated for profile.do, expandListing.do, and results.do

+++ HouseReport/src/main/java/com/skilldistillery/housereport/data/ListingDAO.java
@@ -15,6 +15,6 @@ public interface ListingDAO {
 	boolean delete(Listing listing);
 	Listing update(Listing listing, Address address);
 	List<Listing> listings();
-	int updateRating(int id);
+	double updateRating(int id);
 
 }
